Fix a bug with turn indicator and clean up a lot of code

+++ engine/nine_morris_3d_engine/graphics/opengl/buffer.cpp
@@ -97,7 +97,7 @@ void UniformBuffer::unbind() {
     glBindBuffer(GL_UNIFORM_BUFFER, 0);
 }
 
-void UniformBuffer::set(const void* field_data, unsigned int field_index) {
+void UniformBuffer::set(const void* field_data, size_t field_index) {
     memcpy(data + fields[field_index].offset, field_data, fields[field_index].size);
 }
